+++ config/testconfig.go
@@ -3,8 +3,6 @@ package config
 import (
 	"fmt"
 	"os"
-	"regexp"
-	"strconv"
 	"strings"
 
 	"github.com/joho/godotenv"
@@ -213,119 +211,6 @@ func (c *TestConfig) ReadFromEnvVar() error {
 	}
 
 	return nil
-}
-
-// loadEnvVarGroupedMap scans all environment variables, matches them against
-// a specified pattern, and returns a map of grouped values based on the pattern.
-// The grouping is defined by the first capture group of the regex.
-func loadEnvVarGroupedMap(pattern string) map[string][]string {
-	logger := logging.GetTestLogger(nil)
-	re := regexp.MustCompile(pattern)
-	groupedVars := make(map[string][]string)
-	for _, env := range os.Environ() {
-		pair := strings.SplitN(env, "=", 2)
-		if len(pair) != 2 {
-			continue
-		}
-		key, value := pair[0], pair[1]
-		matches := re.FindStringSubmatch(key)
-		if len(matches) > 1 && value != "" {
-			group := matches[1] // Use the first capture group for grouping
-			groupedVars[group] = append(groupedVars[group], value)
-			logger.Debug().Msgf("Will override test config from env var '%s'", key)
-		}
-	}
-	return groupedVars
-}
-
-func loadEnvVarSingleMap(pattern string) map[string]string {
-	logger := logging.GetTestLogger(nil)
-	re := regexp.MustCompile(pattern)
-	singleVars := make(map[string]string)
-	for _, env := range os.Environ() {
-		pair := strings.SplitN(env, "=", 2)
-		if len(pair) != 2 {
-			continue
-		}
-		key, value := pair[0], pair[1]
-		matches := re.FindStringSubmatch(key)
-		if len(matches) > 1 && value != "" {
-			group := matches[1] // Use the first capture group for grouping
-			singleVars[group] = value
-			logger.Debug().Msgf("Will override test config from env var '%s'", key)
-		}
-	}
-	return singleVars
-}
-
-// Merges a map[string]string with a map[string][]string and returns a new map[string][]string.
-// Elements from the single map are inserted at index 0 in the slice of the new map.
-func mergeMaps(single map[string]string, multi map[string][]string) map[string][]string {
-	newMap := make(map[string][]string)
-
-	// First, copy all elements from the multi map to the new map
-	for key, values := range multi {
-		newMap[key] = make([]string, len(values))
-		copy(newMap[key], values)
-	}
-
-	// Next, insert or prepend the elements from the single map
-	for key, value := range single {
-		if existingValues, exists := newMap[key]; exists {
-			// Prepend the value from the single map
-			newMap[key] = append([]string{value}, existingValues...)
-		} else {
-			// Initialize a new slice if the key does not exist
-			newMap[key] = []string{value}
-		}
-	}
-
-	return newMap
-}
-
-type EnvValueType int
-
-const (
-	String EnvValueType = iota
-	Integer
-	Boolean
-	Float
-)
-
-// readEnvVarValue reads an environment variable and returns the value parsed according to the specified type.
-func readEnvVarValue(envVarName string, valueType EnvValueType) (interface{}, error) {
-	// Get the environment variable value
-	value, isSet := os.LookupEnv(envVarName)
-	if !isSet {
-		return nil, nil
-	}
-	if isSet && value == "" {
-		return "", nil // Return "" if the environment variable is not set
-	}
-
-	// Parse the value according to the specified type
-	switch valueType {
-	case Integer:
-		intVal, err := strconv.Atoi(value)
-		if err != nil {
-			return nil, fmt.Errorf("error converting value to integer: %v", err)
-		}
-		return intVal, nil
-	case Boolean:
-		boolVal, err := strconv.ParseBool(value)
-		if err != nil {
-			return nil, fmt.Errorf("error converting value to boolean: %v", err)
-		}
-		return boolVal, nil
-	case Float:
-		floatVal, err := strconv.ParseFloat(value, 64)
-		if err != nil {
-			return nil, fmt.Errorf("error converting value to float: %v", err)
-		}
-		return floatVal, nil
-	default: // String or unrecognized type
-		return value, nil
-	}
 }
 
 func LoadSecretEnvsFromFiles() error {
